requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
build==1.2.2.post1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-rest-auth==6.0.0
Django
django-allauth==65.1.0
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.4.6
django-filter==24.3
django-modeltranslation==0.19.11
django-redis==5.4.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==30.3.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
Markdown==3.7
packaging==24.1
pillow==10.4.0
pip-tools==7.4.1
prompt_toolkit==3.0.48
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stripe==11.2.0
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
